Extract os-independent Path methods into own class to enable re-use in following commits
