+++ backend/fairTeams.DemoHandling/FTPDemoCollector.cs
@@ -86,7 +86,7 @@ namespace fairTeams.DemoHandling
             using var scope = myServiceProvider.CreateScope();
             var matchRepository = scope.ServiceProvider.GetService<MatchRepository>();
 
-            var existingDemoFileNames = matchRepository.Matches.Select(x => GetFileName(x.Demo.FilePath)).ToList();
+            var existingDemoFileNames = matchRepository.Matches.Select(x => PathExtensions.GetFileName(x.Demo.FilePath)).ToList();
             existingDemoFileNames.AddRange(Directory.EnumerateFiles(Settings.DemoWatchFolder).Where(x => x.EndsWith(".dem")));
             return allDemoFiles.Where(x => !existingDemoFileNames.Contains(x.Name)).ToList();
         }
@@ -151,33 +151,5 @@ namespace fairTeams.DemoHandling
 
             return finishedDemoFiles;
         }
-
-        private static string GetFileName(string path)
-        {
-            var isLinuxPath = IsLinuxPath(path);
-            var systemIsLinux = IsLinux();
-
-            // Path.GetFileName doesn't work for Windows paths running on a Linux system
-            // https://docs.microsoft.com/en-us/dotnet/api/system.io.path.getfilename?view=net-5.0#System_IO_Path_GetFileName_System_String_
-            if (!isLinuxPath && systemIsLinux)
-            {
-                var splitWindowsPath = path.Split("\\").ToList();
-                return splitWindowsPath.Last();
-            }
-
-            return Path.GetFileName(path);
-        }
-
-        private static bool IsLinuxPath(string path)
-        {
-            return path.StartsWith("/");
-        }
-
-        public static bool IsLinux()
-        {
-            int p = (int)Environment.OSVersion.Platform;
-            return (p == 4) || (p == 6) || (p == 128);
-        }
-
     }
 }
